requirements.txt
flask
torch
torchvision
python-dotenv
pytest
datasets
matplotlib
pillow
numpy